+++ R/stratify.R
@@ -139,7 +139,7 @@ stratify <- function(by,
                      species,
                      strata_custom = NULL,
                      combine_species_forms = TRUE,
-                     release = 2024,
+                     release = 2025,
                      sample_data = FALSE,
                      return_omitted = FALSE,
                      quiet = FALSE) {
@@ -239,7 +239,8 @@ stratify <- function(by,
           paste0(.data$country, "-", .data$st_abrev, "-", .data$bcr)))
 
     if(stratify_type == "subset") {
-      meta_strata <- strata_custom
+      meta_strata <- strata_custom %>%
+        sf::st_drop_geometry()
     } else {
       meta_strata <- bbsBayes2::bbs_strata[[stratify_by]]
     }